requirements.txt
absl-py==0.13.0
altair==4.1.0
argon2-cffi==20.1.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
Automat==20.2.0
autopep8==1.5.6
backcall==0.2.0
base58==2.1.0
beautifulsoup4==4.9.3
bleach==3.3.0
blinker==1.4
bokeh
branca==0.5.0
bs4==0.0.1
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==3.0.4
charset-normalizer==2.0.3
clang==5.0
click==7.1.2
cloudpickle==2.0.0
cmdstanpy==0.9.68
colorama==0.4.4
commonmark==0.9.1
constantly==15.1.0
convertdate==2.4.0
cryptography==3.4.7
cssselect==1.1.0
cycler==0.10.0
Cython==0.29.23
decorator==4.4.2
defusedxml==0.7.1
dukpy==0.2.3
emoji==1.5.0
entrypoints==0.3
ephem==4.1.3
filelock==3.6.0
Flask==2.0.1
flatbuffers==1.12
folium
future==0.18.2
gast==0.4.0
gensim==4.1.2
geocoder==1.38.1
gitdb==4.0.7
GitPython==3.1.18
glfw==2.3.0
google-auth==1.34.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
graphviz==0.17
grpcio==1.39.0
gym==0.20.0
h2==3.2.0
h5py==3.1.0
hijri-converter==2.2.3
holidays==0.13
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.6
ImageHash==4.3.1
imageio==2.9.0
imbalanced-learn==0.8.1
importlib-metadata==4.12.0
importlib-resources==5.2.2
incremental==21.3.0
ipykernel==5.5.0
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itemadapter==0.3.0
itemloaders==1.0.4
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
jupyterthemes==0.20.0
keras==2.6.0
Keras-Applications==1.0.8
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keras-resnet==0.2.0
kiwisolver==1.3.1
lesscpy==0.14.0
lockfile==0.12.2
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.0
mistune==0.8.4
mysql-connector-python==8.0.28
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.3.0
numpy==1.19.5
oauthlib==3.1.1
opencv-contrib-python==4.5.3.56
opencv-python==4.5.3.56
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.5
pandocfilters==1.4.3
parsel==1.6.0
parso==0.8.1
pep8==1.7.1
pickleshare==0.7.5
Pillow==8.1.2
plotly==5.6.0
ply==3.11
priority==1.3.0
prometheus-client==0.9.0
prompt-toolkit==3.0.18
Protego==0.1.16
protobuf==3.20.1
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pydeck==0.7.0
PyDispatcher==2.0.5
pydotplus==2.0.2
pyglet==1.5.21
Pygments==2.8.1
PyMeeus==0.5.11
Pympler==1.0.1
pyOpenSSL==20.0.1
PyPAC==0.16.0
pyparsing==2.4.7
pypng==0.20220715.0
PyQRCode==1.2.1
pyrsistent==0.17.3
pystan==2.19.1.1
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.2.0
# pywin32==300
pywinpty==0.5.7
PyYAML==6.0
pyzmq==22.0.3
qrcode==7.1
qtconsole==5.1.0
QtPy==1.9.0
queuelib==1.6.1
ratelim==0.1.6
regex==2021.8.3
requests==2.26.0
requests-file==1.5.1
requests-oauthlib==1.3.0
rich==12.5.1
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.7.0
Scrapy==2.5.0
seaborn==0.11.1
semver==2.13.0
Send2Trash==1.5.0
service-identity==21.1.0
setuptools-git==1.2
six==1.15.0
smart-open==5.2.1
smmap==4.0.0
soupsieve==2.2.1
streamlit
streamlit-folium==0.6.15
streamlit-option-menu==0.3.2
streamlit-player==0.1.5
tabulate==0.8.6
tenacity==8.0.1
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
terminado==0.9.3
testpath==0.4.4
threadpoolctl==2.1.0
tldextract==3.2.1
toml==0.10.2
toolz==0.11.1
torch==1.12.0
tornado==6.1
tqdm==4.62.0
traitlets==5.0.5
Twisted==21.7.0
twisted-iocpsupport==1.0.1
tzdata==2021.1
tzlocal==3.0
ujson==5.1.0
urllib3==1.22
validators==0.18.2
w3lib==1.22.0
watchdog==2.1.5
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wikipedia==1.4.0
wrapt==1.12.1
xgboost==1.4.2
zipp==3.6.0
zope.interface==5.4.0
